マイグレーションファイルのクラス名はマイグレーションファイルのファイル名と同じにしましょう。 それ以外だとクラスが二重あるよっていう全然エラー内容と違うエラーが表示されます。

+++ src/database/migrations/2024_12_05_124530_create_read_receipts_table.php
@@ -4,26 +4,28 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class CreateReadReceiptsTable extends Migration
 {
     /**
      * Run the migrations.
      */
-    public function up()
+    public function up(): void
     {
-        Schema::create('read_receipts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('message_id')->constrained('chat_messages')->onDelete('cascade');
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('read_receipts')) {
+            Schema::create('read_receipts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('message_id')->constrained('chat_messages')->onDelete('cascade');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
      * Reverse the migrations.
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('read_receipts');
     }
-};
+}
